requirements.txt
Django==1.11.20
mysqlclient==1.3.12
djangorestframework==3.6.4
django-cors-headers==2.4.0
djangorestframework-simplejwt==4.1.0
django-widget-tweaks==1.4.1